working on ability to add groups to meeting
