+++ src/actions/gathering.js
@@ -169,6 +169,7 @@ export const createGathering = (formData, history, edit = false) => async (
                 }
             }
         */
+        let junk = window.prompt("WHAT?");
         let Item = formData;
         let obj = {
             operation: 'addGroup',
@@ -196,12 +197,11 @@ export const createGathering = (formData, history, edit = false) => async (
         console.log('that was from actions::gatherings::createGathering');
         // console.log(typeof formData._id);
         // console.log(formData._id.length);
-        if (formData._id.length < 1) {
+        if (formData.meetingId.length < 1) {
             //this is an add, so delete _id and meetingId from formData
-            delete formData._id;
             delete formData.meetingId;
         } else {
-            formData.meetingId = formData._id;
+            formData.meetingId = formData.meetingId;
             //formData._id = '';
         }
         // if(formData._id) formData.push("meetingId", formData._id);
